requirements.txt
langchain
openai
html2text
tiktoken
pinecone-client
unstructured
textloader